chore(registration-react): Move auth-service.js to its own module directory
