+++ applications/registration-react/src/auth/auth-service.js
